YD-433 Fixed activity comment message delete issue
* Break the cyclic relations before performing the delete
* Use explicit delete instead of cascading
* Brought the cycle-breaking for the self-reference of the thread head under the same umbrella
* Added integration tests
